# Cache class names and resolvable types