--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleA/ConfigA.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleA/ConfigA.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleA;
+
+use Gacela\Framework\AbstractConfig;
+
+final class ConfigA extends AbstractConfig
+{
+    public function getConfigValue(): string
+    {
+        return $this->get('config-key');
+    }
+}