--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleA/DepProvA.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleA/DepProvA.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleA;
+
+use Gacela\Framework\AbstractDependencyProvider;
+use Gacela\Framework\Container\Container;
+
+final class DepProvA extends AbstractDependencyProvider
+{
+    public function provideModuleDependencies(Container $container): void
+    {
+        $container->set('provided-dependency', 'dependency-value');
+    }
+}