--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleA/Facade.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleA/Facade.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleA;
+
+use Gacela\Framework\AbstractFacade;
+
+/**
+ * @method FactoryA getFactory()
+ */
+final class Facade extends AbstractFacade
+{
+    public function loadGacelaCacheFile(): array
+    {
+        return $this->getFactory()->getArrayConfigAndProvidedDependency();
+    }
+}