--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleA/FactoryA.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleA/FactoryA.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleA;
+
+use Gacela\Framework\AbstractFactory;
+use GacelaTest\Fixtures\StringValueInterface;
+
+/**
+ * @method ConfigA getConfig()
+ */
+final class FactoryA extends AbstractFactory
+{
+    private StringValueInterface $stringValue;
+
+    public function __construct(StringValueInterface $stringValue)
+    {
+        $this->stringValue = $stringValue;
+    }
+
+    public function getArrayConfigAndProvidedDependency(): array
+    {
+        return [
+            'config-key' => $this->getConfig()->getConfigValue(),
+            'string-value' => $this->stringValue->value(),
+            'provided-dependency' => $this->getProvidedDependency('provided-dependency'),
+        ];
+    }
+}