--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleB/ConfigB.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleB/ConfigB.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleB;
+
+use Gacela\Framework\AbstractConfig;
+
+final class ConfigB extends AbstractConfig
+{
+    public function getConfigValue(): string
+    {
+        return $this->get('config-key');
+    }
+}