--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleB/DepProvB.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleB/DepProvB.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleB;
+
+use Gacela\Framework\AbstractDependencyProvider;
+use Gacela\Framework\Container\Container;
+
+final class DepProvB extends AbstractDependencyProvider
+{
+    public function provideModuleDependencies(Container $container): void
+    {
+        $container->set('provided-dependency', 'dependency-value');
+    }
+}