--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleB/Facade.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleB/Facade.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleB;
+
+use Gacela\Framework\AbstractFacade;
+
+/**
+ * @method FactoryB getFactory()
+ */
+final class Facade extends AbstractFacade
+{
+    public function loadGacelaCacheFile(): array
+    {
+        return $this->getFactory()->getArrayConfigAndProvidedDependency();
+    }
+}