--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleC/ConfigC.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleC/ConfigC.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleC;
+
+use Gacela\Framework\AbstractConfig;
+
+final class ConfigC extends AbstractConfig
+{
+    public function getConfigValue(): string
+    {
+        return $this->get('config-key');
+    }
+}