--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleD/ConfigD.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleD/ConfigD.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleD;
+
+use Gacela\Framework\AbstractConfig;
+
+final class ConfigD extends AbstractConfig
+{
+    public function getConfigValue(): string
+    {
+        return $this->get('config-key');
+    }
+}