--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleE/ConfigE.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleE/ConfigE.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleE;
+
+use Gacela\Framework\AbstractConfig;
+
+final class ConfigE extends AbstractConfig
+{
+    public function getConfigValue(): string
+    {
+        return $this->get('config-key');
+    }
+}