--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleE/DepProvE.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleE/DepProvE.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleE;
+
+use Gacela\Framework\AbstractDependencyProvider;
+use Gacela\Framework\Container\Container;
+
+final class DepProvE extends AbstractDependencyProvider
+{
+    public function provideModuleDependencies(Container $container): void
+    {
+        $container->set('provided-dependency', 'dependency-value');
+    }
+}