--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleF/Config.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleF/Config.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Benchmark\Framework\ClassResolver\ClassNameCache\ModuleF;
+
+use Gacela\Framework\AbstractConfig;
+
+final class Config extends AbstractConfig
+{
+    public function getConfigValue(): string
+    {
+        return $this->get('config-key');
+    }
+}