--- a/tests/Benchmark/Framework/ClassResolver/ClassNameCache/config/default.php
+++ b/tests/Benchmark/Framework/ClassResolver/ClassNameCache/config/default.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'config-key' => 'config-value',
+];