… composer.json
… tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleF/Facade.php
… tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleF/Factory.php
… tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleG/Config.php
… tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleG/DependencyProvider.php
… tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleG/Facade.php
… tests/Benchmark/Framework/ClassResolver/ClassNameCache/ModuleG/Factory.php
… tests/Unit/Framework/ClassResolver/ClassNameFinderTest.php
… tests/Unit/Framework/ClassResolver/InMemoryClassNameCache.php