## 📚 Description

Create a `.cache` file with the "classInfo->cacheKey" and its "resolvable candidate" as a php array, so it can be `required` and used on the fly without losing performance.

## 🔖 Changes

- Extract the cached classes from `ClassNameFinder` and create its own separated class for it: `ClassNameCache`.
- The `ClassNameCache` will start with the content of the `'.gacela-class-names.cache'` and it will update the content of the cache file every-time it find and resolves a new classInfo.
- Added some benchmark tests to prove the speed improvement when using the cache vs without cache.

> Most of the added lines are the dummy modules for the benchmark tests.

## 🖼️ Screenshot

The same code executed without and with cache:

- Without cache:  3.9 ms
- With cache: 263.8 μs

> 1 ms = 1,000,000 μs

<img width="1534" alt="Screenshot 2022-06-04 at 21 23 19" src="https://user-images.githubusercontent.com/5256287/172022665-f34077dd-b0c7-4556-a495-fe29e7c6aaa9.png">